requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
motor>=3.0.0
pymongo>=4.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.0.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
jinja2>=3.0.0
python-dotenv>=0.19.0
email-validator>=1.3.0